[FIX]: code readability
