--- src/helpers/enums.ts
@@ -1,6 +1,0 @@
-
-export enum Gender {
-  MALE = 'MALE',
-  FEMALE = 'FEMALE',
-  OTHER = 'OTHER',
-}
