+++ src/helpers/index.ts
@@ -5,7 +5,6 @@ export * from './strings';
 export * from './time';
 export * from './utils';
 export * from './responsive';
-export * from './enums';
 
 export {
   DeviceUtils
