--- src/views/Home/HomeScreen.style.ts
@@ -1,17 +1,0 @@
-
-import {StyleSheet} from 'react-native';
-
-const homeScreenStyles = (theme: IAppTheme) =>
-  StyleSheet.create({
-    container: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-      backgroundColor: theme.colors.background,
-    },
-    text: {
-      color: theme.colors.red,
-    }
-  });
-
-export default homeScreenStyles;
