--- src/views/Register/RegisterScreen.style.ts
@@ -1,14 +1,0 @@
-
-import {StyleSheet} from 'react-native';
-
-const registerScreenStyles = (theme: IAppTheme) =>
-  StyleSheet.create({
-    container: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-      backgroundColor: theme.colors.background,
-    },
-  });
-
-export default registerScreenStyles;
