--- src/views/Splash/SplashScreen.style.ts
@@ -1,30 +1,0 @@
-
-import {rpWidth} from '@app/helpers';
-import {StyleSheet, ViewStyle} from 'react-native';
-import {ImageStyle} from 'react-native-fast-image';
-
-interface Style {
-  container: ViewStyle;
-  version: ViewStyle;
-  splsh_image: ImageStyle;
-}
-
-const splashScreenStyles = () =>
-  StyleSheet.create<Style>({
-    container: {
-      justifyContent: 'center',
-      alignItems: 'center'
-    },
-    version: {
-      position: 'absolute',
-      bottom: 5,
-      right: 5
-    },
-    splsh_image: {
-      height: rpWidth(200),
-      width: rpWidth(200),
-      borderRadius: rpWidth(100)
-    }
-  });
-
-export default splashScreenStyles;
