--- src/views/Webview/WebviewScreen.style.ts
@@ -1,14 +1,0 @@
-
-import {StyleSheet} from 'react-native';
-
-const webviewScreenStyles = (theme: IAppTheme) =>
-  StyleSheet.create({
-    container: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-      backgroundColor: theme.colors.background,
-    },
-  });
-
-export default webviewScreenStyles;
